some changes on projectDao & project
