+++ src/model/item/Item.java
@@ -6,6 +6,8 @@
 package model.item;
 
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleIntegerProperty;
+import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 import model.servicetype.ServiceType;
 
@@ -14,8 +16,59 @@ import model.servicetype.ServiceType;
  * @author bek
  */
 public class Item {
+    
     private IntegerProperty id;
+    private IntegerProperty projectId;
     private StringProperty name;
-    private IntegerProperty numberOfWords;
+    private IntegerProperty numberWords;
+    private StringProperty createdDate;
     private ServiceType service;
+  
+    public Item() {
+        this.id = new SimpleIntegerProperty();
+        this.projectId = new SimpleIntegerProperty();
+        this.name = new SimpleStringProperty();
+        this.numberWords = new SimpleIntegerProperty();
+        this.createdDate = new SimpleStringProperty();
+    }
+
+    public IntegerProperty getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id.set(id);
+    }
+    
+    public IntegerProperty getProjectId() {
+        return projectId;
+    }
+
+    public void setProjectId(Integer projectId) {
+        this.projectId.set(projectId);
+    }
+    
+    public StringProperty getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name.set(name);
+    }
+    
+    public IntegerProperty getNumberWords() {
+        return numberWords;
+    }
+
+    public void setNumberWords(Integer numberWords) {
+        this.numberWords.set(numberWords);
+    }
+    
+    public StringProperty getCreatedDate() {
+        return createdDate;
+    }
+
+    public void setCreatedDate(StringProperty createdDate) {
+        this.createdDate = createdDate;
+    }
 }
