+++ src/model/item/ItemDAO.java
@@ -1,4 +1,4 @@
-package model;
+package model.item;
 
 import util.DBUtil;
 import javafx.collections.FXCollections;
@@ -6,53 +6,49 @@ import javafx.collections.ObservableList;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import model.project.Project;
 
 /**
  *
  * @author Khandaa
  */
-public class ProjectDAO {
+public class ItemDAO {
   
-    public static Project find(String id) throws SQLException, ClassNotFoundException {
-        String selectStmt = "SELECT * FROM project WHERE id="+id;
+    public static Item find(Integer id) throws SQLException, ClassNotFoundException {
+        String selectStmt = "SELECT * FROM item WHERE id="+id;
         try {
             ResultSet rs = DBUtil.dbExecuteQuery(selectStmt);
           
-            Project p = null;
+            Item i = null;
             if (rs.next()) {
-                p = new Project();
-                //p.setId(rs.getInt("ID"));
-                p.setTitle(rs.getString("TITLE"));
-                p.setAmount(rs.getDouble("AMOUNT"));
-                p.setCurrency(rs.getString("CURRENCY"));
-                //p.setVendorId(rs.getInt("VENDOR_ID"));
-                //p.setCreatedDate(new SimpleIntegerProperty(rs.getDate("CREATED_DATE")));
+                i = new Item();
+                i.setId(rs.getInt("ID"));
+                i.setProjectId(rs.getInt("PROJECT_ID"));
+                i.setName(rs.getString("NAME"));
+                i.setNumberWords(rs.getInt("NB_WORDS"));
             }
-            return p;
+            return i;
         } catch (SQLException e) {
-            System.out.println("While searching a project with " + 
+            System.out.println("While searching an item with " + 
                     id + " id, an error occurred: " + e);
             throw e;
         }
     }
 
-    public static ObservableList<Project> list() throws SQLException, 
+    public static ObservableList<Item> list() throws SQLException, 
             ClassNotFoundException {
-        String selectStmt = "SELECT * FROM project";
+        String selectStmt = "SELECT * FROM item";
         try {
             ResultSet rs = DBUtil.dbExecuteQuery(selectStmt);
-            ObservableList<Project> list = FXCollections.observableArrayList();
+            ObservableList<Item> list = FXCollections.observableArrayList();
 
-            Project p;
+            Item i;
             while (rs.next()) {
-                p = new Project();
-                //p.setId(rs.getInt("ID"));
-                p.setTitle(rs.getString("TITLE"));
-                p.setAmount(rs.getDouble("AMOUNT"));
-                p.setCurrency(rs.getString("CURRENCY"));
-                //p.setVendorId(rs.getInt("VENDOR_ID"));              
-                list.add(p);
+                i = new Item();
+                i.setId(rs.getInt("ID"));
+                i.setProjectId(rs.getInt("PROJECT_ID"));
+                i.setName(rs.getString("NAME"));
+                i.setNumberWords(rs.getInt("NB_WORDS"));
+                list.add(i);
             }
             return list;
         } catch (SQLException e) {
@@ -61,13 +57,13 @@ public class ProjectDAO {
         }
     }
     
-    public static void insert(String title, Double amount, String currency, Integer vendorId) 
+    public static void insert(Integer projectId, String name, Integer numberWords) 
             throws SQLException, ClassNotFoundException {
         String updateStmt =
             "BEGIN\n" +
-                "INSERT INTO project\n" +
-                "(TITLE, AMOUNT, CURRENCY, VENDOR_ID)\n" +
-                "VALUES('"+title+"', "+amount+", '"+currency+"', "+vendorId+");\n" +
+                "INSERT INTO item\n" +
+                "(PROJECT_ID, NAME, NB_WORDS)\n" +
+                "VALUES('"+projectId+"', '"+name+"', "+numberWords+");\n" +
                 "END;";
 
         try {
@@ -78,15 +74,14 @@ public class ProjectDAO {
         }
     }
    
-    public static void update (String id, String title, Double amount, String currency, Integer vendorId) 
+    public static void update (String id, Integer projectId, String name, Integer numberWords) 
             throws SQLException, ClassNotFoundException {
         String updateStmt =
             "BEGIN\n" +
-                "   UPDATE project\n" +
-                "      SET title = '" + title + "'\n" +
-                "      SET amount = " + amount + "\n" +
-                "      SET currency = '" + currency + "'\n" +
-                "      SET VENDOR_ID = " + vendorId + "\n" +
+                "   UPDATE item\n" +
+                "      SET PROJECT_ID = " + projectId + "\n" +
+                "      SET NAME = '" + name + "'\n" +
+                "      SET NB_WORDS = " + numberWords + "\n" +
                 "    WHERE ID = " + id + ";\n" +
                 "   COMMIT;\n" +
                 "END;";
@@ -101,7 +96,7 @@ public class ProjectDAO {
     public static void delete(String id) throws SQLException, ClassNotFoundException {
         String updateStmt =
             "BEGIN\n" +
-                "   DELETE FROM project\n" +
+                "   DELETE FROM item\n" +
                 "         WHERE id ="+ id +";\n" +
                 "   COMMIT;\n" +
                 "END;";
